Fix mechanical compiler warnings: static access, redundant casts, serialVersionUID
- Qualify static field/constant access by type name instead of by an
  instance expression (PrintPreview, SearchField, DialogSymbolize).
- Remove redundant casts flagged by the compiler across the codebase
  (all confirmed no-ops by javac's own type analysis).
- Add serialVersionUID to Serializable Swing classes missing it.

No behavioral change; verified with a clean -Xlint:all recompile
before and after (309 warnings down to 137, all remaining ones left
for a follow-up pass since they need case-by-case review: non-transient
serializable fields, raw types, this-escape, lossy conversions).

+++ src/fidocadj/circuit/controllers/EditorActions.java
@@ -250,11 +250,11 @@ public class EditorActions
 
         // Calculate a reasonable tolerance. If it is too small, we ensure
         // that it is rounded up to 2.
-        int toll=(int)Math.round(cs.unmapXnosnap(x+sel_tolerance)
+        int toll=Math.round(cs.unmapXnosnap(x+sel_tolerance)
             -cs.unmapXnosnap(x));
         if (toll<2) { toll=2; }
-        selectPrimitive((int)Math.round(cs.unmapXnosnap(x)),
-            (int)Math.round(cs.unmapYnosnap(y)),
+        selectPrimitive(Math.round(cs.unmapXnosnap(x)),
+            Math.round(cs.unmapYnosnap(y)),
             toll, toggle);
     }
